requirements.txt
blobfile==3.0.0
celery==5.2.7
Django==3.2.12
lmdb==1.6.2
matplotlib==3.7.1
numpy==1.25.0
pandas==1.4.1
Pillow==9.4.0
prettytable==3.16.0
python-decouple==3.8
setuptools==75.8.0
tqdm==4.67.1
